% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{element n is the sum of element n−1 and the absolute value of the difference between elements n−2 and n−3 divided by two.}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{factor}

\item{n}{integer}
}
\value{

}
\description{
element n is the sum of element n−1 and the absolute value of the difference between elements n−2 and n−3 divided by two.
}
