% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numSeqPlot.R
\name{numSeqPlot}
\alias{numSeqPlot}
\title{Plot a graphic from myseq_n}
\usage{
numSeqPlot(nums)
}
\arguments{
\item{nums}{The input from the same package called function myseq_n}
}
\value{
This function should return a line plot of the output values for the different values of n.
}
\description{
The first three columns are the values of the three numerics to be input to function myseq_n() and the fourth column is the positive integer n for the sequence to be generated.
This function should return a line plot of the output values for the different values of n.
}
