% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04ptwyunting-package.R
\docType{package}
\name{hw04ptwyunting-package}
\alias{hw04ptwyunting}
\alias{hw04ptwyunting-package}
\title{hw04ptwyunting: My package for homework week 4 by stat-613}
\description{
A recursive formula of three sequences with proper graphic
}
\author{
\strong{Maintainer}: Yunting Chiu \email{yc6705a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
