% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numSeqPlot.R
\name{numSeqPlot}
\alias{numSeqPlot}
\title{Plot a graphic from myseq_n}
\usage{
numSeqPlot(nums)
}
\arguments{
\item{nums}{The input from the same package called function myseq_n}
}
\value{
This function should return a line plot of the output values for the different values of n.
}
\description{
The first three columns are the values of the three numeric to be input from function myseq_n() and the fourth column is the positive integer n for the sequence to be generated.
This function should return a line plot of the output values for the different values of n.
}
\examples{
# testing the function as data frame
my_data <- tibble::tribble(
~x, ~y, ~z, ~n,
2,4,3,3,
2,4,3,4,
2,4,3,5,
2,4,3,6,
2,4,3,7,
2,4,3,8,
2,4,3,9,
2,4,3,10,
2,4,3,12)
numSeqPlot(my_data)
}
